Update the query to extract the HLT key based on the run number
Update the query to extract the HLT key based on the run number to use
the OMS database tables rather than the WBM database tables.

Note that the OMS team does not guarantee that the structure and content
of the database tables are stable, and strongly advice to use the
API-based interface for any new tool.

See https://its.cern.ch/jira/browse/CMSOMS-272 for more information.