change default compression settings of OutputModules
https://its.cern.ch/jira/browse/CMSHLT-2651 changes the default compression settings of the HLT OutputModules, i.e.

 compression_level = cms.untracked.int32(3),
 compression_algorithm = cms.untracked.string("ZSTD")

This commit updates the constructor of the OutputModule class in order to use the new default values.
This way, the new default does not need to be set manually every time a new OutputModule is created
(if non-default values are needed, one will have to edit the new `OutputModule` manually).